Move some console-related message types into the console code.
